+++ matlab/src/main/java/net/maswag/falcaun/simulink/SimulinkModel.java
@@ -1,4 +1,4 @@
-package net.maswag.falcaun;
+package net.maswag.falcaun.simulink;
 
 import com.mathworks.engine.EngineException;
 import com.mathworks.engine.MatlabEngine;
@@ -7,6 +7,10 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import net.automatalib.word.Word;
+import net.maswag.falcaun.Signal;
+import net.maswag.falcaun.TimeMeasure;
+import net.maswag.falcaun.ValueWithTime;
+
 import org.apache.commons.lang3.ArrayUtils;
 
 import javax.annotation.Nonnull;
