+++ matlab/src/test/java/net/maswag/falcaun/simulink/AutotransExample.java
@@ -1,7 +1,12 @@
-package net.maswag.falcaun;
+package net.maswag.falcaun.simulink;
 
 import lombok.Getter;
+import net.maswag.falcaun.AdaptiveSTLUpdater;
+import net.maswag.falcaun.IOSignalPiece;
+import net.maswag.falcaun.NumericSULMapper;
+import net.maswag.falcaun.NumericSULVerifier;
 import net.maswag.falcaun.parser.SimpleSignalMapper;
+import net.maswag.falcaun.simulink.SimulinkSULVerifier;
 
 import org.apache.commons.collections4.CollectionUtils;
