+++ matlab/src/test/java/net/maswag/falcaun/simulink/SimulinkSULMapperTest.java
@@ -1,6 +1,8 @@
-package net.maswag.falcaun;
+package net.maswag.falcaun.simulink;
 
 import net.automatalib.word.Word;
+import net.maswag.falcaun.IOSignalPiece;
+import net.maswag.falcaun.NumericSULMapper;
 import net.maswag.falcaun.parser.SimpleSignalMapper;
 
 import org.junit.jupiter.api.BeforeEach;
