+++ matlab/src/test/java/net/maswag/falcaun/simulink/SimulinkSULVerifierTest.java
@@ -1,10 +1,20 @@
-package net.maswag.falcaun;
+package net.maswag.falcaun.simulink;
 
 import org.junit.jupiter.api.*;
 
+import net.maswag.falcaun.AdaptiveSTLList;
+import net.maswag.falcaun.AdaptiveSTLUpdater;
+import net.maswag.falcaun.ArgParser;
+import net.maswag.falcaun.IOSignalPiece;
+import net.maswag.falcaun.NumericSULMapper;
+import net.maswag.falcaun.StaticSTLList;
+import net.maswag.falcaun.StopDisprovedEQOracle;
+import net.maswag.falcaun.ArgParser.GASelectionKind;
+import net.maswag.falcaun.StopDisprovedEQOracle.StaticLTLList;
 import net.maswag.falcaun.parser.STLFactory;
 import net.maswag.falcaun.parser.SimpleSignalMapper;
 import net.maswag.falcaun.parser.TemporalLogic;
+import net.maswag.falcaun.simulink.SimulinkSULVerifier;
 
 import java.util.*;
 import java.util.function.Function;
