+++ matlab/src/test/java/net/maswag/falcaun/simulink/SimulinkSignalTest.java
@@ -1,6 +1,8 @@
-package net.maswag.falcaun;
+package net.maswag.falcaun.simulink;
 
 import net.automatalib.word.Word;
+import net.maswag.falcaun.Signal;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
